+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 #### v0.3.0
-* [breaking] In envconfig_derive: rename function envconfig() -> derive().
+* [breaking] In envconfig_derive: rename function envconfig() -> derive()
+* [breaking] Remove init_or_die function from Envconfig trait
+* Add better documentation to the crate
 
 #### v0.2.0 - 2018-09-13
 * [breaking] Use derive macro instead of macro_rules